Refactor: change directory structure
