+++ src/quri_parts_oqtopus/backend/__init__.py
@@ -1,25 +1,32 @@
-from .sampling import (  # noqa: I001
+from quri_parts_oqtopus.backend.jobs.sampling import (  # noqa: I001
     OqtopusSamplingBackend,
 )
 
-from .data.sampling import (
+from quri_parts_oqtopus.models.jobs.sampling import (
     OqtopusSamplingJob,
+)
+
+from quri_parts_oqtopus.models.jobs.results.sampling import (
     OqtopusSamplingResult,
 )
 
-from .device import OqtopusDevice, OqtopusDeviceBackend
-from .estimation import (
+from .devices.device import OqtopusDeviceBackend
+from quri_parts_oqtopus.backend.jobs.estimation import (
     OqtopusEstimationBackend,
 )
 
-from .data.estimation import (
+from quri_parts_oqtopus.models.jobs.estimation import (
     OqtopusEstimationJob,
+)
+from quri_parts_oqtopus.models.jobs.results.estimation import (
     OqtopusEstimationResult,
 )
 
-from .sse import OqtopusSseBackend
+from quri_parts_oqtopus.backend.jobs.sse import OqtopusSseBackend
 
 from .config import OqtopusConfig
+
+from quri_parts_oqtopus.models.devices.device import OqtopusDevice
 
 __all__ = [
     "OqtopusConfig",
