+++ src/quri_parts_oqtopus/backend/base.py
