+++ src/quri_parts_oqtopus/backend/jobs/sampling.py
@@ -89,8 +89,8 @@ from quri_parts.openqasm.circuit import convert_to_qasm_str
 from quri_parts_oqtopus.backend.config import (
     OqtopusConfig,
 )
-from quri_parts_oqtopus.backend.data.sampling import OqtopusSamplingJob
-from quri_parts_oqtopus.backend.job_backend_base import OqtopusJobBackendBase
+from quri_parts_oqtopus.backend.jobs.base import OqtopusJobBackendBase
+from quri_parts_oqtopus.models.jobs.sampling import OqtopusSamplingJob
 from quri_parts_oqtopus.rest import (
     JobsSubmitJobInfo,
     JobsSubmitJobRequest,
