+++ src/quri_parts_oqtopus/models/utils.py
